Second proof for topological invariance etale site
This is the proof of the equivalence of etale sites under a
universal homeomorphism. The first proof uses the very delicate
arguments about comparison of sites in various forms earlier
in the chapter on etale cohomology. This material is kinda
hard to read and maybe a bit nonstandard. The new proof
is just like the proof in SGA 1 and uses the descent for
etale morphisms along surjective integral morphisms + the
topological invariance for thickenings... In fact in some
sense the new proof may very well be harder, but it seems
at least easier to break it up into parts which have meaning...