Fix error in proof
Thanks to Lena Min Ji

The error was that we did not check carefully
that M-regular sequences in I would map to
regular sequences in the localization at the
chosen prime. And in fact with the choice of
prime in the original argument this would in
general NOT be the case, whence the argument was
wrong. We corrected it by choosing a prime in
the support of M/IM.