Proof descirption of j_{F, *}
Thanks to Xu Gao who writes

"I'm a reader of stack project. I found that in 04H0, the author states
that "In order to prove that this works the introduction of Hom-sheaves
is desirable, hence we postpone this to a later time." But the notion of
Hom-sheaves has already introduced in 04TQ although the adjunction of
product and Hom is missing. So I make this supplement, what I have
changed are 04TQ and  04H0. Do you think this change is reasonable?"