Fix mistake in more-algebra
Not a fatal mistake by any means and not bad at all, but annoying to
fix. Namely, it seems that we need to use the following when we prove
that our property is inherited by henselization or strict
henselization: given

k ---> k' ---> R

where the first arrow is a separable algebraic extension of fields and
where k ---> R has property P, then k' ---> R has property P as well.
Now if k'/k is finite, then this follows from (A) and (B). But in the
case of an infinite extension it does not seem to formally follow from
(A) and (B). Thus we added it as a new property and we checked this
property in all of the cases...