Bound on the l-torsion in Pic num type
Finally! What I enjoy about this way of proving things is that you get a
linear bound on the bad primes l. Of course, this is useless since if
you want to get the best bound, you should probably look at semistable
reduction of the Jacobian and then it suffices to look at l=3.