Try to correct a composition of morphisms
I suggest at least this correction since the line after the correction doesn't follow otherwise. Technically, some identifications have been made without comment in this proof, i.e. between the image and coimage of $f$ (because abelian) and between the kernel of $g$ and the image of $f$ (because exact). And making these identifications has meant that some of the sequences of morphisms have gotten screwed up.