Fix error in the proof of Gabber's lemma
Thanks to Brian Conrad for pointing it out
http://stacks.math.columbia.edu/tag/0AIX#comment-1947

The error (due to me) was that twice we used that the product of open
ideals is open in the ring A^\wedge without justification (and surely it
isn't true). What is true is that the closure of the product of two open
ideals is open and that is enough for the proof to go through.