Existence of global object gerbe if H^2 = 0
This lemma is just enough for the application we have in mind later. For
the Stacks project it perhaps makes sense to discuss this kind of result
in complete generality after first dicussing a bit more the relationship
between hypercoverings and cohomology as in the chapter on simplicial
spaces...

This should be on the todo list.