Connected components and base field extension
Thanks to Ofer Gabber who writes

Let me now answer the question in Tag 04PZ.  Let Tbar be a connected
component of X_K and T the  connected component of X containing the
image of Tbar;  we may replace X by T with some subscheme structure so
WMA that X is connected. Let A=O(X) and consider L=the ind-étale part
of A, which is a field as A does not have nontrivial idempotents. For
a finite separable extension  L' of L we may view X as an L-scheme and
base change it to X' over L' ; then by the above the ring of global
functions on X' is identified with A tensored over L with L', and its
largest ind-étale subalgebra
(over k, L or L')  is L',  in particular X' is connected.  Thus by
Tag 0389   X is geometrically connected over L.  Now Tbar maps to a
connected component Spec(E) of Spec(K tensored over k with L)
(work with the reduced subscheme structures) ,so Tbar is contained in
X"=X base changed from L to E viewed as a subschema of X_K, but X" is
connected, so equal to Tbar, and surjects to X.