Fix snafu in flat.tex
The proof of a small part of lemma-weak-bourbaki-pre
was erroneous... The problem was in the sentence

"Let $z \in M \otimes_R N$ be an element whose annihilator in $S$
has radical $\mathfrak q$."

In general weakly associated primes don't automatically have
this property, only after you localize does it become true.
To work around it we invented a trick where you choose a z
such that \mathfrak q is minimal over its annihilator and
then you look at what happens after you map z up into a
suitably chosen overring (using the machinery of RG).

If you look at the proof of the corresponding fact in RG
then you don't find any explanation of how to deal with this
whatsoever, so this could mean there is a much simpler way
to deal with it. Anyway, I am quite convinced the current
argument is correct, so I will leave it like this for now.