Existence of universal flattenings (spaces)
Exactly the same as for schemes. Note the strange work around at the
beginning of the proof of theorem. This is only because the chapter
comes before the bootstrap chapter... or because we haven't added a
discussion of descending separated locally quasi-finite morphisms along
fppf coverings in the setting of algebraic spaces yet.