Strong formal effectiveness => openness versality
Actually not hard to prove and you only need to check on first order
thickenings.

Bhargav's example from c31011d shows that strong formal effectiveness
does not always hold. But what about strong formal effectiveness
where the thickenings are always first order (between any two, not
just between consecutive indices)? This is the only thing needed for
the argument here, so it would be nice if it was true for algebraic
stacks in general.