Temper assertions of level of difficulty
Thanks to Brian Conrad who wrote:

"I just briefly glanced through the section on sst reduction for curves,
and have a few minor comments:

1. In 48.1 you mention Gieseker?s proof and say he proves the full
result.  But strictly speaking, doesn?t his method only apply as written
in the equicharacteristic case?  Also, the way 48.1 is written seemed at
first sight to be suggesting that Chapter 48 would likewise avoid
desingulaization; maybe it is worth clarifying there that the method of
Artin-Winters (which I have never looked at) does use Lipman?s
desingularization result.   The case-checking in section 48.5 is quite
something!

2.  When you say in 48.1 that sst reduction for abelian varieties seems
a lot harder, is that because of the role of the theory of Neron models
(whose construction is quite non-trivial, etc.)?  If so, maybe it is
worth saying this.  The proof given in the case of curves as in Ch. 48
is invoking resolution for surfaces and various things on dualizing
complexes.  When all is done, is that not an equally difficult task? I
haven?t read the treatment in SP, so maybe there are simplifications
incorporated (and in the end I suppose it is a judgement call; the saga
of birational group laws over a base is very hard stuff?).

3. In Remark 48.18.2 at the end of 48.18 you refer to [DM], but strictly
speaking [DM] was written assuming the residue field is alg. closed (see
pp. 90?91, where k being alg. closed is used in several of the
calculations in the harder direction that semistable reduction for J ==>
semistable reduction for C), so in effect perfect residue field. This is
a harmless hypothesis on the residue field for the sake of proving sst
reduction for curves without restriction on the residue field (since
over the henselization we can reach algebraic closure of the residue
field through a rising tower of finite separable extensions with e=1),
as is implicit in Artin-Winters too with their starting assumption on
the residue field. But for 48.18.2 this hypothesis on the residue field
is less evidently harmless. Note for example that Saito?s paper which
you mention assumes algebraically closed residue field (or
?equivalently? for these purposes, perfect residue field).  Is it known
by another method that this equivalence as stated in 48.18.2 is valid
for general residue fields?"

My reply to this was the following:

"OK, I looked over your comments. I agree with all of them and I will
make changes accordingly in a week or so.

About Gieseker's proof: At the time GIT wasn't available over Z, but
now his method just works over Z. The key input is just that a smooth
projective curve over a field embedded.with a high power of an ample
line bundle gives a GIT stable points of a Hilbert scheme and that all
GIT semi-stable points of this Hilbert scheme correspond to nodal
curves. These are both geometric statements you can prove over
algebraically closed fields; then automatically you obtain ss
reduction.

Yes, desingularization for surfaces is a key input in Artin-Winters.
Yes, the resolution part needs a tiny bit of duality -- it could be
developed from scratch in a few pages if one wanted to. But
interestingly, the actual argument in the SP for ss reduction in
Artin-Winters avoids using the dualizing sheaf and relative duality.

Some of the papers in the literature say something like: "There is an
immediate problem to the case where the residue field is algebraically
closed." but do not give the reduction itself (I think this is true
for the Saito paper for example and for Artin-Winters)."