More fallout change definition in 5e6f086
Again in the lemma we changed here
(a) the original formulation was actually insufficient for the use
of lemma-colim-works-over-collection, and
(b) there is a sharper formulation of this lemma when working with
quasi-compact and quasi-separated objects

TODO: Define quasi-separated objects to sites.tex

Discussion: you can try to say that a quasi-compact object U of
of a site is quasi-separated if for all pairs of quasi-compact
objects V, W over U one has that W \times_U V is quasi-compact.
But there are at least two problems with this as far as I can see:
(1) maybe the fibre product W \times_U V does not exist
(2) maybe there are very few quasi-compact objects over U
You can get around (1) by considering the associated sheaves...
however I think it would be more natural to assume there are lots
of quasi-compact objects, for example, require that every covering
of U can be refined by a covering with quasi-compact objects) and
then problem (1) goes away too (as the morphisms in a covering
are representable by definition).