More notational changes
TODO: The section on proper hypercovering has to be rewritten

We should introduce morphisms between simplicial sites and
then we should state that the morphism

c : LC_{qc}/U ---> U

satisfies descent for abelian sheaves. We have already proved
this actually, but we haven't stated it.

Then this proves Lemma \ref{lemma-compute-via-proper-hypercovering}
directly by looking at the commutative square and using the
already proven descent for

a : (LC_{qc}/U)_{total} ---> LC_{qc}/X

Finally, we deduce Lemma \ref{lemma-spectral-sequence-proper-hypercovering}
as a consequence of this rather than the other way around.

Doing the same for Ra_* will give a better proof of
Lemma \ref{lemma-cohomological-descent-for-proper-hypercovering}
as well.