BBD simplicial glueing lemma for D(Ab)
Seems very solid already on the first pass.

Interestingly you don't need to use vanishing of all negative exts
for the question for morphisms, but you do seem to need all negative
exts to vanish in order to construct objects that are locally given.
At least that is what happens with this method...