Fix a small mistake
The problem was in the lemma producing a hypercovering consisting of
weakly contractible objects. In the proof we did not check the maps
compose correctly and in general they do not.

So now we rely on the very general construction of hypercovers
given an ample supply of objects and then we afterwards take
disjoint unions of objects to get weakly contractible ones.

This stuff is just a matter of bookkeeping but you still have
to do the bookkeeping correctly.