Spell out duality theory a bit more
Direct the reader to the results computing what f^! is like
in specific cases. One missing case is where one has a smooth
morphism (and not just a smooth and proper morphism). It could
easily be deduced from the properties already esthablished
for the f^! functors in the same way we do for the smooth and
proper case. In fact, one could do this more generally for
locally complete intersection morphisms

TODO: Elucidate relative dualizing complex for lci morphisms