Compare qc and Zar on locally compact spaces
Hopefully this is the last time I have to go through this.
The original argument with the Cech spectral seqence was
also fine, but it really needed much more explanation as
to why it worked. I think with the current approach the
argument carries over for etale vs ph topologies and for
etale vs fppf topologies... but we will see...