Comparing small and big etale sites
Improve the exposition and fix an error

The error was that the restriction from modules on the big etale
site of a scheme to the small etale site does not send injective
modules to injective modules. Made a counterexample. This error
does not affect anything else in the Stacks project. It occurs
in another location which we fix in the next commit.