Stacks Project additional text
Thanks to Burt Totaro who writes:

"I would like to add a bit of explanation to the Stacks Project.

Namely, Tag 05B5 is a very useful fact (that smoothness
descends under fppf coverings), but it is a little bit hidden.
It is not mentioned in the introduction to the section,
or in later sections where this fact is relevant.
So I suggest adding a few references to this fact,
and a related example on smoothness of quotient stacks.
See details below."