Additivity not needed
Thanks to Dajano Tossici who writes:

"...just a remark about 010M. It is seems to me (up to a trivial mistake)
that in Lemma 12.7.1 (ii),(iii),(iv) the hypothesis 'additive' is not
necessary, i.e. I think that if the functor sends, for instance, a short
left exact sequence in a short left exact sequence then it is additive.
Clearly a priori F(0) \neq 0, so the map 0\to F(A) is just the unique
morphism from the initial object to F(A) (and not F(0)\to F(A))."