Generalize section on infinitesimal deformations
We want to use the definitions even when the base scheme is not
necessarily Noetherian.

We also removed the requirement in (RS*) that the affine schemes in
question map into an affine open of the base scheme. This is never
needed when one actually verifies a given stack has (RS*) and in
fact this more general notion is always a consequence of the
original restricted one.