Move section on dualizing modules to obsolete
Try to make it clear that we don't need to use the relative dualizing
module for models in order to prove semistable reduction for curves.

Of course we used a little bit of duality in the proof of resolution of
singularities for surface singularities, for example in the statement
and proof of Grauert-Riemenschneider vanishing.