Final fallout error fixed in f6e7e84c
This is quite annoying, but the original argument relied on the
erroneous lemma from the commit given above...

Question: Is it actually just true that the pullback of a K-flat complex
along a morphism of ringed topoi is K-flat? Seems to me this is possibly
true but I don't immediately see how to prove it.

Also, it does not seem terribly important.