A few fixes
Thanks to Ofer Gabber (new mistakes are mine) who writes:

"Concerning the notion of locally ringed topos, which was sometimes given
imprecisely in references  ,e.g. in the book of Monique  Hakim Chap.III
Def.2.3, in spite of the attention to 0 stalks this imprecision
propagated to Tag 04ES lemma 18.39.1, where (1) and (3) hold for the
zero ring, but condition (2) with n=0 implies that if O(U) is 0 than U
admits an empty covering family,  so the conditions  are not equivalent.

In Tag 0ASN Lemma 15.98.2, condition (2) implies that R is not zero
since by convention 0 is not local,but not (1),(3), so should add that R
is nonzero.

 In Lemma 37.25.10 Tag 0ATA the intersection  of the V(I)´s should be
the union; also A should not be a field, otherwise the set {A} is a
counterexample, the mistake is  that in the end one needs that f is in
the maximal ideal, so observe that under the corrected statement there
must be a proper ideal I in S ,take f to be a nonzero element of the
maximal ideal,  was  OK in my scan.

 The following are pedantic objections to the notion of site. In
Def.7.6.2 tag 00VH ,in condition (1) it is not clear what is the one
point index set  for V---->U,  and in condition (3) it is not clear
whether we require it to hold for some choice of the fibred products (of
U_i and V over U) or for all choices. In Remark 7.47.4 Tag 00ZF it is
annoying to have Cov(C) a subset of the power set of Arrow(C), because
for the empty set one cannot determine the target object (think of the
stupid site with a discrete category where for certain objects one
admits the empty covering family) ; so may consider P(Arrow(C))xOb(C)
instead...."