Clarify construction (strict) henselization
Thanks to Minseon Shin who writes:

"I have a question regarding the henselization
construction Tag 04GN, in the argument showing that the
category of pairs (S,\mathfrak{q}) satisfies condition
(3) (i.e. has coequalizers) in the definition of a
filtered category (002V). In order to show that
R \to S'' is etale, isn't it necessary to use at some
point the fact that etale morphisms satisfy the
cancellation property 00U7? Looking at the diagram, it
seems we need to know that \varphi, \psi, and the
diagonal S' \otimes_{R} S' \to S' are etale.

I thought this was worth pointing out since this part
seems to be what fails when we attempt the same
construction for "etale" replaced by "finitely presented
flat" (i.e. to construct the "fppf stalk" of a local
ring)."