Simplify proof key lemma in chow
Large set of changes in order to simplify the proof of the key lemma
used in proofs of properties for intersecting with effective Cartier
divisors and capping with c_1 of invertible modules.

Probably the original proof was a bit too original. It is still there
in an appendix, but the current proof can be read within an hour by
those familiar with standard (Noetherian) commutative algebra. The
original proof required checking commutativity of many diagrams and
checking many signs.

We also added more explanation of what is happening in the introduction
as well as a comment on Milnor K-theory giving the reference to Kato's
paper and how it is more general.