Update algebra.tex
As suggested in the todo-list we add a little "bit more to the section on the snake lemma in algebra.tex". 
In particular, we add a remark and a new reference (Algebra, Serge Lang) before the lemma itself, and we improve the logical structure of the proof by making more explicit where one uses the assumptions, while keeping the proof not verbose.