transfinite devissage of modules
Thanks to Dario Weissmann who writes

Dear Johan,

so there was an index-typo in lemma-direct-sum-devissage. And for the
injectivity of f_{\beta} I think we needed to calculate the image of
f_{\beta} not just show that $f$ is surjective. Thought it would be
more difficult to explain what I meant in a post than just changing it.