Uniquess of arrow in Lemma 28.50.5
Thanks to Fei XU who writes

Dear Stacks-Project authors:

I would like to emphasis the uniqueness of morphism $X_2' \rightarrow
X_1′$ in Lemma 28.50.5 in Stacks Project instead of "a canonical
diagram" in the last sentence of this lemma. The uniqueness follows from
the uniqueness of Lemma 28.50.4 (2). Then these two lemmas are
compatible. "Canonical" (usually means functoriality) is not quite clear
to me.