Error in proof 0A8F
Thanks to Eike Lau who writes

Tag 0A8F
In the proof, I think we don’t have H(X) = lim H(X_n) as stated
but only a surjective map H(X) \to lim H(X_n).
This is sufficient for the rest of the argument without any change.
(The kernel of the surjective map is lim^1 H(X_n), but this is not
important.)