Fix idiotic error
Thanks to anonymous for pointing this out

Let me just say that there is another way to check this lemma if there
are enough points. Namely, in this case the stalks of the lower shriek
at a point of C is the direct sum of the points of C/U lying over this
point and the result follows immediately. In fact, this argument is the
reason for trying to prove the lemma in the first place.

Of course this doesn't excuse us for having a totally bogus argument!