Gabber's criterion for henselianness...
Thanks to Kestutis Cesnavicius who writes:
---------------

... a proof for Gabber's criterion of Henselianess in the
attached files. There is no good reference for this (Gabber's sketch
consists of only 6 lines!), so I think it would be good if this could
be incorporated into the Stacks Project. I've indicated the appropriate
places where this could be inserted in the attached pdf.

Also, I think it would be very nice to have a section like
https://stacks.math.columbia.edu/tag/09XD but for "Zariski pairs,"
i.e., for those (A, I) with I \subset \rad(A). For instance,
https://stacks.math.columbia.edu/tag/09XF really belongs to such a section.