Fix for material in previous commit
A blow up of a non-Noetherian scheme need not be of finite presentation
even if the center is cut out by a finite type ideal. Thus we introduce
"almost blow up squares". This actually makes the whole thing work
better because almost blow up squares are preserved under base change
and in the Noetherian case a blow up square is an almost blow up square.

For more details, please just read the code.