Blow up squares and ph topology
Skeleton section.

It is NOT possible to characterize ph sheaves in terms of fppf coverings
and blow up squares. The problem is that there are proper surjective
morphisms which cannot be made flat by a finite stratification of the
affine base... Example

Base is the spectrum of A = Z[x_0, x_1, x_2, x_3, ...]
Source is the Proj of the graded ring A[S, T]/I where I is generated by
the elements

x_0(S - T), x_1(S^2 - T^2), x_2(S^4 - T^4), x_3(S^8 - T^8), ...

Over the stratum where x_0 = ... = x_{e - 1} = 0 but where x_e is
nonzero we get a degree 2^e cover...