Split out arguments in proof smooth base change
I decided to leave in the very long proof because it seems that for some
people this may be easier to read than the version with all the
references to the lemmas building up the proof.

Also added: some more cases where the base change maps are isomorphisms
on sheaves.