Bit the bullet on comparisons
Thanks to Bhargav Bhatt and Carl Lian

This new material is going to replace the 4 times or so repeated
arguments of comparing different cohomologies. It also fixes a snafu in
the proof of a lemma in each of these which was pointed out by Bhargav
during a very clear seminar talk by Carl Lian. Namely, in the proof of

0DCV,sites-cohomology-lemma-V-implies-C
0DDP,etale-cohomology-lemma-V-implies-C-etale-fppf
0DDZ,etale-cohomology-lemma-V-implies-C-etale-ph
0EWB,etale-cohomology-lemma-V-implies-C-etale-h

we don't correctly check what happens to the spectral sequences and the
argument isn't complete. What is missing is an additional argument in
the last degree (n) which is done in the current commit in the general
setup...