A mild modification to Beauville-Laslo
Thanks to Kestutis Cesnavicus who writes:
------------------

I attach a minor edit to the section https://stacks.math.columbia.edu/tag/0BNI
on the Beauville--Laszlo theorem --- could you include it in the
Stacks Project? The goal is to generalize the section to more general maps
R ---> R', not just those that arise from the f-adic completion. This makes
this material better suited for an eventual nonaffine statement.

I tried to be careful, hopefully I did not introduce any errors. The same
proofs worked in the more general situation with only cosmetic changes
(I went through the arguments in detail).