Minor fix in "less naive" stackification
The construction claims that the functor Cat -> Sets that maps
a category to the disjoint union of its objects and arrows
commutes with limits which is not correct: For two categories
C and D, the objects resp. arrows of C x D are just the product
of corresponding objects resp. arrows, but the product of
disjoint unions of objects and arrows has "mixed terms" like
Ob(C) x Arrows(D).

I think that this issue can be fixed by replacing the forgetful
functor by (Ob, Arrows) : Cat -> Sets x Sets: It is still
faithful, reflects isomorphisms, actually commutes with limits
(as it "is" the right adjoint of the left Kan extension of
F: {o, a} -> Cat with F(o)=[0] and F(a)=[1] along the Yoneda
embedding {o, a} -> Sets^{{o,a}^op} ~ Sets x Sets) and commutes
with directed colimits (which can be checked by verifying the Segal
condition for the componentwise colimit using the fact that
directed colimits commute with pullbacks).

A minor issue is that Tag 00YR ("Sheaves of algebraic structures")
cannot be used directly because the target of the forgetful functor
is not Sets, but sheafification should still commute with this
forgetful functor as (co)limits in Sets x Sets are computed
componentwise.