Update to "Divided power algebras"
Thanks to Burt Totaro who writes:

----------------------

I attach a proposed revision of the chapter
"Divided power algebras" in the Stacks Project
(as the Latex file "dpa.tex"). The changes
are only in the section "Tate resolutions".
I also attach an addition to the bibliography
as the file "dpa.bib".

The point of the change is to improve one result:
namely, we show that the Tor algebra Tor_*^R(S,T) has divided powers
for arbitrary homomorphisms of commutative rings
R -> S and R -> T. There is no need for Noetherian
or finite type assumptions. This seems important enough
that the general statement should appear
in the Stacks Project.

My changes may not require any changes to the tags. I added one
remark, but otherwise I just added a few lines
in statements or proofs to point out that we can use
resolutions with infinitely many generators
in each degree, if we like. Then all the proofs work as before.

The earliest paper I found that explicitly observes
that Tor has divided powers in full generality
is Avramov-Halperin's "Through the looking glass" (1986).
So I refer to that paper in the text,
and I attach a bibliography entry for that paper as "dpa.bib".

-----------------------