Better handling f_! for loc quasi-finite
Thanks to Will Sawin for suggesting to construct f_! in this manner.

The text doesn't check commutativity all the diagrams in great detail.
To do this would require a better language of talking about functors
which are "pinned" in the sense that the stalks of their images are
fixed to be something specific...