Lemmas on weakly etale morphisms
Thanks to Shishir Agrawal who writes

--------------

I've attached a tex file with a couple of lemmas about weakly etale
morphisms, likely of the "well known to experts" variety. Perhaps they
might fit in towards the end of the section on weakly etale morphisms in
the "More on Morphisms" chapter of the Stacks Project...? (There'd be a
"forward reference" to a result about algebraic spaces for which I
couldn't find a scheme theoretic analog stated in the Stacks Project. I
don't think this causes any circularities, but I don't know if it would
be stylistically acceptable.)

---------------

I've replaced the forward reference to one in the chapter on descent.