The Tate map
Found a way to prove a global isomorphism

det(L_{Y/X}) ---> omega_{Y/X}

for lqf lci morphisms of locally Noetherian schemes where in the proof
you NEVER have to change coordinates. The key is to find suitable versal
deformations and realize that the isomorphism is clearly unique for
etale morphisms... The idea for this comes from the paper by Garel cited
in the section.