Some cleanups + a link between two traces
The original disussion of the trace used for the purely inseparable case
in the chapter on resolution of surface singularities could now be
replaced by a reference to the traces on differential forms. However, it
seems worthwhile to leave in place the construction in this particularly
easy case which is a key step in the resolution process and at the same
time is much easier than the general trace construction.