Fix error
Thanks to Bogdan Zavyalov

The error was to silently assume that X is dense in any compactification
of X but in the definition of compactifications earlier in the Stacks
project we didn't require this. This was a bit hard to fix because we
then in later lemmas used the flexibility afforded by the more general
definition of a compactification.

I fixed this by leaving the definition of a compactification alone and
by very carefully arguing that inside the category of all
compacfifications of a given scheme X the ones where X is dense is an
initial category.

I also much clarified the proof of Lemma 0B6T to boot

I also checked all other uses of compactifications in the Stacks project
and the lemmas I fixed are the only ones using compactifications in the
wrong manner...

Interestingly in the case of defining etale cohomology with proper
supports we don't make the same mistake because there we have a stronger
lemma about lower shriek functors which makes the thing work more
smoothly.