Rough outline arguments on fully faithful functors
Try as much as possible to move arguments into general categorial
nonsense... This works but it doesn't recover the functor as a
Fourier-Mukai functor, it just tells you there is a "sibling" which is a
Fourier-Mukai functor. Since this is enough for the main application I
have in mind, I am going to try not to think about improvements for now.