Fix missing arguments in proof OCXU
Thanks to 李一笑 who writes:

"The proof implicitly used that any dense subset of a noetherian
topological space contains a countable dense subset. Maybe this should
be made explicit and be formulated into a lemma."

After asking more 李一笑 also pointed out that

"What's more, it may not be the case that $E$ is infinite, for example
if there is an element of $E$ whose closure contains $u_{0}$. This case
should also be considered in 0CXS."

The first problem is addressed in this commit and the previous one
(7f41eb8f but I could only prove the existence of the countable dense
subset in the case of Noetherian schemes). The second problem was
addressed in 3e7633b0