Compactly supported cohomology coherent modules
Just the beginning. This is not really what Deligne wanted to do because
we don't check the well definedness of Rf_! ab ovo but instead we use
the fact that we've already esthablished the existence of f^! (and the
fact that it is well defined etc). But I judge that the checks needed to
verify the well definedness of Rf_! from first principles is pretty much
just as hard as what we had to do for f^! and we've already done it.

But please complain to me anyway if you want!