Fix silly mistake in Kunneth discussion
Just bad! The mistake was in Remark with latex label remark-silly which
we have turned into a Section with latex label section-silly.

To fix it was more interesting: you can formulate a kunneth type result
for complexes made out of quasi-coherent module whose differentials
are differential operators of finite order. We will apply this to
the de Rham complex (whose differentials are differential operators of
order 1) in the next commit!