Proof d = 0 on H^d(X, Omega^{d - 1})
It seems the trick with the splitting of R\Gamma(X, O_X) into A and a
perfect object of tor amplitude in [1, d] is kind of necessary? Don't
know for sure.

I'm quite interested in places in the literature where this is proven in
different ways. Let me know if you have any.