Fix a mistake in the proof of 01VR
Thanks to Luke
https://stacks.math.columbia.edu/tag/01VR#comment-5799
https://stacks.math.columbia.edu/tag/01VR#comment-5813
https://stacks.math.columbia.edu/tag/01VR#comment-5816

I am still dissatisfied with the proof as it is written now. But I am
dissatisfied with all of the chapter on Proj and this is just an example
of a case where if we had done a better job on Proj then this would be
easier to say.

Of course, I also think the Proj stuff is just impossible to write up in
a good way!