Algebraization rig-smooth in the G-ring case
Finally! The proof is actually much better than before as we now do not
worry about the cotangent complex for the nearby algebra and instead
worry only about very course properties of it. Thus the approximation
needs only approximate the generators of the ideal and the relations
between them and nothing else...