an error in 0F40
Thanks to 李一笑 who writes

--------------------------

(in the paragraph starting with defining V_{12}) The morphism p_1 may
not be an isomorphism over an open neighbourhood of Z_{1,2}, and is
only a bijective closed immersion in general, since we only assumed
U_1 \cap U_2 to be set-theoretically dense in U.

To fix the problem we may instead define V_{12} = V_1 \times_U V_2,
and X_{12} the scheme-theoretic image of V_{12} \to X_1 \times_S X_2.
Then the remaining parts of the proof need not be changed.

p.s. If you want to fix the original proof,
the equality p_1^{-1} V_1 = p_2^{-1} V_2 defining V_{12} is true,
but it is not obvious, and may require proof.

---------------------------

All of this is correct (I think). I have fixed the proof by taking
the first suggestion. All mistakes are mine!