A bit of uniformity: revision
Although the arguments as given before this commit are correct, they
don't prove the result in the form that we actually use them later in
the chapter on algebraization of rig-smooth algebras. Sigh! We still
have to fix the proof of this algebraization result to account for the
changes we just made, but that should be easy and quick!