Down the rabbit hole: rig-everything
Got caught up in infinite regression having to do with the availablility
of basic notions on rig-descent... Namely, I want to

(a) show that if we have X -> Y -> Z with X -> Y rig-smooth +
rig-surjective then Y is rig-smooth over Z if and only if X is so

(b) this requires me to check that the sequence of naive cotangent
complexes in the rig setting has a 0 on the left of the 6-term sequence

(c) this in turn requires me to check that a rig-smooth morphism is sort
of a rig-flat rig-complete-intersection morphism

(d) this made me look into the notion of rig-flatness and this made me
realize that it isn't actually as straightforward as you might think
because actually it's easy to make mistakes. In fact I think the BL
papers don't do enough checks to see that their notion even makes
sense. Of course, it does make sense if you always take sufficiently
small neighbourhoods on your formal scheme... but I'm not sure if they
literally say this correctly.

ARGHHHH!!!