Fix proof of existence separable solution
Thanks very much to Takeshi Saito who found the error and proposed the
solution. First he observed that we used multiplicativity of inseparble
degrees for field extensions which aren't algebraic and hence the proof
cannot work. He found a counterexample to a part of the proof. Then he
proposed the solution as follows:

"I think that the lemma itself is fine.
You just need to take an approximation of K2 by a separable extension."

All remaining mistakes are mine of course.