Fact on finite modules over Noetherian rings
Interesting that this was never explicitly stated but of course used all
the time. It sort of shows that the setup of the Stacks project sort of
works: we start somewhere in the "middle" between really basic algebra
and geometry and work "upwards". Still it doesn't hurt to also work on
fixing missing initial material.