Tensor functor gives morphisms
Thanks to Jarod Alper and Max Lieblich

We prove that a right exact tensor functor Coh(X) ---> Coh(Y) comes
from a morphism Y ---> X if X, Y are Noetherian and X is an algebraic
stack with affine diagonal. The proof is short because we use a trick to
show that a certain ring map is universally injective and then we use
descent for universally injective ring maps.