First higher chow group + affine bundles
We have all the ingredients completely contained in the Stacks project,
but I am going to leave it as a sketch of proof for the moment. The
lemma we generalized in commit 2f248527 is exactly what is needed to
show the desired properties of flat pullback on CH^m_k(X, 1).