Fun improvement of Orlov's proposition due to C&S
Thanks to Noah Olander

Noah points out that in the paper by Canonano and Stellari they are able
to drop the vanishing of positive Exts from the result of Orlov.
Analyzing their proof, Noah figured out how to minimally change our
current proof and get the same result (which you can find in this
commit).