Valuative criteria - DVRs essentially of finite type
Thanks to Jochen Heinloth who writes:

"In the application part of the joint article with Dan and Jarod on the
existence of good moduli spaces we needed to reduce our valuative
criteria for DVRs to DVRs essentially of finte type - we are probably
not alone with this issue.

As a referee kindly pointed out, this reduction was not addressed
carefully. During the revision process I noticed that I could not find
valuative criteria of this type explicitly in the stacks project,
Section 32.15/tag 0CM1 seemed to be the natural place. Maybe I just
didn't search well enough, but I was wondering whether that might be
useful to add (we will put something in the revision of our paper ). One
statement is attached, but you will know a better one."