Add reference to SGA 6
Thanks to Juhani Bonsdorff who writes:

"Remark 10.69.8 (tag 061T) states that both Koszul-regularity and H_1-regularity were introduced by T. Kabele in his 1971 paper.

However, while the name "Koszul-regular" and the variant concept of H_1-regularity indeed seem to be originating here, (as Kabele also vaguely points out)
the definition of Koszul-regularity is actually given already at the beginning of Exposé VII of SGA 6 and is presumably due to Grothendieck or Berthelot.
It is furthermore stated on p. 4 of Exp. VII that this new definition (i.e. Koszul-regularity) was intended to replace the old one in EGA IV.4 (which had just gone to printing), at least in the definition of regular embeddings. It is the one they use for local complete intersection morphisms to make

I think it would make sense to both point out the actual origin of Koszul-regularity and to give a reference to SGA 6."